Rewrote `README.md` as an instructor landing page: a plain... (+13 more)
## 2026-09-21
- Rewrote `README.md` as an instructor landing page: a plain-prose About paragraph, a status
  line, a badge row, a centered WeBWorK Question Library hero, and three product statements (practi...
- Applied the screenshot-recovery cleanup map: replay staging now requires its manifest;
  publication receipts and atlases explicitly cover the active public, Instructor, and Student role...
- Kept the Live Demo launcher single-purpose by removing its acceptance-only
  `--without-live-demo` mode and opt-out acceptance fixture. The lower-level Cargo installation-dat...
- Published a fresh 97-capture screenshot corpus from a clean Live Demo, including 33 Instructor
  captures, 56 Student captures, all 18 unanswered and 18 saved-answer question-type laptop/phone c...
- Restored the semantic `webwork_` prefix on Instructor generated-example screenshots. Viewport
  variants now use explicit role folders instead of redundant `_laptop` and `_phone` filename suffi...
- Organized viewport-qualified screenshots under role viewport folders (`laptop/`, `tablet/`,
  `phone/`, and `square/`) while keeping genuinely single-view artifacts at the role root. The Stud...
- Hardened the Rust verification front door so generated Question-ID contracts are checked rather
  than rewritten; removed temporary publication-response coupling from the screenshot scenarios; an...
- Completed the pre-production identity cleanup: Rust and TypeScript now distinguish
  `PublishedQuestionId` from `QuestionPoolId`, exact pins are `PublishedQuestionRevisionTuple`, and...
- Re-ran the complete screenshot corpus from a clean Live Demo after the identity cleanup. The
  capture now uses Student identities for the HOTSPOT workflow and disposable assessments for its u...
- Recorded an independent screenshot-recovery audit for the coding manager. It reviews the
  cross-layer repair batch, Question-identity naming, Student Question-format coverage, and accepta...
## 2026-09-20
- Restored the full acceptance run: Course Banner PostgreSQL/MinIO persistence now uses
  explicit PostgreSQL domain and enum boundaries, transaction-safe object timestamps, explicit work...
- Recorded the one-time `plpgsql_check` PostgreSQL 17 diagnostic in

+++ schemas/base_schema/50_functions/assessment_operations.sql
@@ -176,7 +176,7 @@ SET search_path = pg_catalog, ple_api, ple_data AS $$
            assessment.assessment_status,
            assessment.origin_kind,
            blueprint.blueprint_course_id,
-           assessment.source_blueprint_revision_number,
+           assessment.source_blueprint_revision_number::bigint,
            assessment.source_blueprint_assessment_id,
            assessment.assessment_type,
            policy.assessment_title,
@@ -420,4 +420,3 @@ SET search_path = pg_catalog, ple_api, ple_data AS $$
     ) AS result
       JOIN ple_data.assessment ON assessment.assessment_id = result.assessment_id
 $$;
-
